Working on changes in diagnostic logging and tracking down todo app bug.

+++ lib/Runtime/Debug/TTSerialize.h
@@ -448,7 +448,7 @@ namespace TTD
     {
         int32 SourceLine;
         uint32 EventTime;
-        uint32 TimeHash;
+        uint64 TimeHash;
     };
 
     bool IsDiagnosticOriginInformationValid(const DiagnosticOrigin& info);
